Minor changes in services

+++ src/service/task.service.ts
@@ -1,4 +1,4 @@
-import { Task } from "../models/task.model";
+import { TaskStatus,Task } from "@prisma/client";
 import { TaskRepository } from "../repository/task.repository";
 
 export class TaskService {
@@ -9,27 +9,63 @@ export class TaskService {
     }
 
   createTask = async (data: any): Promise<Task> => {
-    const createdTask = await this._taskRepository.createTask(data);
+   try {
+    const task: Task = {
+      id: data.id,
+      title: data.title,
+      description: data.description,
+      taskstatus: TaskStatus.TODO,
+      userId: data.userId,
+      createdAt: new Date(),
+    };
+    const createdTask = await this._taskRepository.createTask(task);
     return createdTask;
+  }
+  catch(error){
+    throw new Error(`Task with ID ${data.id} not found`);
+  }
   };
 
   getTasks = async (limit:number, offset:number): Promise<Task[]> => {
-    const tasks = this._taskRepository.getTasks(limit, offset);
-    return tasks;
+    try {
+      const tasks = await this._taskRepository.getTasks(limit, offset);
+      return tasks;
+    } catch (error) {
+      throw new Error(`Task with ID ${limit} not found`);
+    }
   };
 
   getTask = async (id: number): Promise<Task> => {
-    const task = this._taskRepository.getTask(id);
-    return task;
+   try {
+     const task = await this._taskRepository.getTask(id);
+     return task;
+   } catch (error) {
+    throw new Error(`Task with ID ${id} not found`);
+   }
 };
   updateTask = async (data: any): Promise<Task> => {
-    const updatedTask = this._taskRepository.updateTask(data);
-    return updatedTask;
+   try {
+     const existingTask = await this._taskRepository.getTask(data.id);
+     if (!existingTask) {
+       throw new Error(`Task with ID ${data.id} not found`);
+     }
+     const updatedTask = await this._taskRepository.updateTask(data);
+     return updatedTask;
+   } catch (error) {
+      throw new Error(`Task with ID ${data.id} not found`);
+   }
 };
-  deleteTask = async (id: number): Promise<Task> => {
-
-    const deletedTask = this._taskRepository.deleteTask(id);
-    return deletedTask;
+  deleteTask = async (id: number, userId:number): Promise<Task> => {
+ try {
+  const existingTask = await this._taskRepository.getTask(id);
+  if (!existingTask) {
+    throw new Error(`Task with ID ${id} not found`);
+  }
+     const deletedTask =await this._taskRepository.deleteTask(id);
+     return deletedTask;
+ } catch (error) {
+    throw new Error(`Task with ID ${id} not found`);
+ }
 };
 }
